requirements.txt
python-telegram-bot==21.0.1
python-dotenv==1.1.0
google-api-python-client==2.122.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pytest==8.1.1
ruff==0.3.3
